fixed Plane, object order and triangle issues

+++ asg3/VS2013_Solution/Assignment3/utils.h
@@ -1,6 +1,8 @@
 #include "..\glm\glm.hpp"
 
 double getMagnitude(glm::vec3 inputVec){
-	double temp = (double)powf(inputVec.x, 2.0) + powf(inputVec.y, 2.0) + powf(inputVec.z, 2.0);
-	return (double)sqrtf(temp);
+	double temp = (double)(powf(inputVec.x, 2.0) + powf(inputVec.y, 2.0) + powf(inputVec.z, 2.0));
+	
+	return (double)sqrtf((float)temp);
+
 }
